+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 0.3.0
 
+- fix: iOS build support
 - feat: auth repository
+- feat: phone auth service
 - feat: auth bloc
 - fix: flutter 2.10 android support